test: replace bad unit stubs with correct integration + unit tests per guideline
- SettingsView: integration test (real AsyncInjectable reactive state, DOM assertions only)
- useApiKeyManagement: unit test (field-status state machine, no mock interaction assertions)
- Remove with-setup.ts helper (was only used by the deleted bad tests)
- Fix file naming to .integration.test.ts / .unit.test.ts per TEST_CLASSIFICATION.md
